requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
lxml==5.3.0
openpyxl==3.1.5
pydantic==2.9.2
python-dotenv==1.0.1
